created an api route for inserting one product...
to the database and integrated the api request on form submit in
AddItemModal Component
