+++ src/types/IProdForm.ts
@@ -1,0 +1,5 @@
+export default interface IProdForm {
+  prod_name: string;
+  prod_price: number | "";
+  prod_quantity: number | "";
+}
